+++ src/utils/__tests__/get-embedded-app-url.test.ts
@@ -48,7 +48,7 @@ describe('getEmbeddedAppUrl', () => {
     } as http.IncomingMessage;
 
     expect(() => getEmbeddedAppUrl(req)).toThrow(
-      ShopifyErrors.InvalidRequestError,
+      ShopifyErrors.InvalidHostError,
     );
   });